add tests: CountMoreThanTotal, RequestIsCorrectly, CityIsUnSupported
